Move deposit API to it's own dir
