+++ src/api/deposit/DepositApiMock.ts
@@ -7,9 +7,9 @@ import { CONTRACT, RECEIPT, createFlux } from '../../../test/data'
 
 import { Receipt, TxOptionalParams } from 'types'
 import { waitAndSendReceipt } from 'utils/mock'
-import { DepositApi } from './DepositApi'
+import { DepositApi, PendingFlux } from './DepositApi'
 import { Erc20Api } from 'api/erc20/Erc20Api'
-import { BalanceState, PendingFlux } from './ExchangeApi'
+import { BalanceState } from '../exchange/ExchangeApi'
 
 export interface BalancesByUserAndToken {
   [userAddress: string]: { [tokenAddress: string]: BalanceState }
