+++ src/api/index.ts
@@ -5,8 +5,8 @@ import { TokenListApiImpl, TokenList } from './tokenList/TokenListApi'
 import { TokenListApiMock } from './tokenList/TokenListApiMock'
 import { Erc20ApiMock } from './erc20/Erc20ApiMock'
 import { Erc20ApiImpl, Erc20Api } from './erc20/Erc20Api'
-import { DepositApiMock } from './exchange/DepositApiMock'
-import { DepositApiImpl, DepositApi } from './exchange/DepositApi'
+import { DepositApiMock } from './deposit/DepositApiMock'
+import { DepositApiImpl, DepositApi } from './deposit/DepositApi'
 import { ExchangeApiImpl, ExchangeApi } from './exchange/ExchangeApi'
 import { ExchangeApiMock } from './exchange/ExchangeApiMock'
 import {
